backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.1.0
supabase==2.15.1
openai==1.75.0
httpx==0.28.1
